// source/packages/@aws-lza/interfaces/aws-cloudformation/create-stack-policy.ts
/**
 *  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 *
 *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
 *  with the License. A copy of the License is located at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 *  or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES
 *  OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions
 *  and limitations under the License.
 */

import { IModuleCommonParameter } from '../../common/resources';

/**
 * Stack Policy module handler parameter
 */
export interface IStackPolicyHandlerParameter extends IModuleCommonParameter {
  readonly acceleratorPrefix: string;
  readonly accountIds: string[];
  readonly enabled: boolean;
  readonly regions: string[];
  readonly managementAccountAccessRole: string;
  readonly protectedTypes: string[];
}

export interface IStackPolicyModule {
  /**
   * Handler function to manage stack policy
   *
   * @param props {@link IStackPolicyHandlerParameter}
   * @returns status string
   */
  handler(props: IStackPolicyHandlerParameter): Promise<string>;
}